requirements.txt
vtkplotlib>=2.0.0
numpy-stl>=2.17.1